[AIRFLOW-3462] Move TaskReschedule out of models.py
