+++ tests/sensors/test_base_sensor.py
@@ -23,7 +23,8 @@ from mock import Mock
 from airflow import DAG, configuration, settings
 from airflow.exceptions import (AirflowSensorTimeout, AirflowException,
                                 AirflowRescheduleException)
-from airflow.models import DagRun, TaskInstance, TaskReschedule
+from airflow.models import DagRun, TaskInstance
+from airflow.models.taskreschedule import TaskReschedule
 from airflow.operators.dummy_operator import DummyOperator
 from airflow.sensors.base_sensor_operator import BaseSensorOperator
 from airflow.ti_deps.deps.ready_to_reschedule import ReadyToRescheduleDep
